Modif

+++ src/components/DeleteTodo.js
@@ -1,9 +1,32 @@
 import { useTodosDispatch } from "../context/TodosDispatchContext"
+import { useIsMounted } from "../hooks/useIsMounted"
 
 const DeleteTodo = ({ todo }) => {
   const dispatch = useTodosDispatch()
+  const isMounted = useIsMounted()
   const deleteTodo = () => {
-    dispatch({ type: "DELETE", payload: todo })
+    fetch(`${process.env.REACT_APP_API_URL}/todos/${todo.id}`, {
+      method: "DELETE",
+      headers: {
+        "Content-type": "application/json",
+      },
+    })
+    .then((response) => {
+      if (!response.ok) {
+        throw new Error(`Something went wrong: ${response.statusText}`)
+      }
+      return response.json
+    })
+    .then (() => {
+      if (isMounted) {
+        dispatch({ type: "DELETE", payload: todo })
+      }
+    })
+    .catch((error) => {
+      if (isMounted) {
+        dispatch({ type: "FETCH_FAILURE", payload: error.message })
+      }
+    })
   }
   return (
     <button
